pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "A collection of various endpoints"
authors = [
    { name = "Luke100000", email = "jonas.peche@aon.at" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.1",
    "aiofile>=3.9.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "audioread>=3.0.1",
    "cachetools>=5.5.0",
    "databases>=0.9.0",
    "diskcache>=5.6.3",
    "dynaconf>=3.2.11",
    "fastapi[standard]>=0.118.0",
    "fastapi-cache2>=0.2.2",
    "fusion-solar-py>=0.0.28",
    "horde-openai-proxy>=0.1.0",
    "itchio>=1.1.3",
    "markdownify>=1.2.0",
    "minecraft-recipe-renderer>=0.1.0",
    "numpy>=2.0.2",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.3",
    "patreon>=0.5.0",
    "pillow>=11.0.0",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydub>=0.25.1",
    "pygithub>=2.8.1",
    "pyrate-limiter>=3.9.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=4.2.0",
    "requests>=2.32.3",
    "torch>=2.8.0",
    "transformers>=4",
    "coqui-tts>=0.27.2",
    "piper-tts>=1.3.0",
]

[dependency-groups]
dash = [
    "a2wsgi>=1.10.10",
    "ciso8601>=2.3.3",
    "dash-bootstrap-components>=2.0.4",
    "dash-bootstrap-templates==1.1.0",
]
dev = [
    "jupyter>=1.1.1",
    "pre-commit>=4.3.0",
    "ruff>=0.13.3",
]
llms = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langchain-mistralai>=0.2.12",
    "langchain-openai>=0.3.34",
    "litellm>=1.77.5",
    "openai>=2.1.0",
]
tts = [
    "cutlet>=0.5.0",
    "mecab-python3>=1.0.10",
    "unidic-lite>=1.0.8",
]

[tool.uv.sources]
horde-openai-proxy = { git = "https://github.com/Luke100000/horde-openai-proxy.git" }
minecraft-recipe-renderer = { git = "https://github.com/Luke100000/minecraft-recipe-renderer.git" }
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
